requirements.txt
google-cloud-logging
google-cloud-storage
google-cloud-bigquery
requests
oauth2client